-- 임시로 PlayerStart를 배치하고 DefaultPawnClass = nullptr과 PioneerController->SetViewTargetWithBlend(PioneerManager->GetWorldViewCamera());를 주석처리하여 볼 수 있게 함. -- Packet.h에 cInfoOfSpaceShip 클래스를 정의하고 SpaceShip에 네트워크용 변수를 추가한 뒤, SetInfoOfSpaceShip(...), GetInfoOfSpaceShip() 함수로 cInfoOfSpaceShip와 데이터를 교환할 수 있도록 함.

+++ Game/Source/Game/GameMode/OnlineGameMode.h
@@ -139,7 +139,7 @@ private:
 	void GetScoreBoard();
 	void RecvScoreBoard();
 	/** 게임서버 연결 확인용 */
-	int CountRecvOfGameServer; 
+	float CountRecvOfGameServer; 
 
 	/////////////////////////////////////////////////
 	// 동기화
